+++ library-gen/Stratosphere/ResourceProperties/ElasticLoadBalancingV2TargetGroupTargetDescription.hs
@@ -21,7 +21,8 @@ import Stratosphere.Values
 -- convenient constructor.
 data ElasticLoadBalancingV2TargetGroupTargetDescription =
   ElasticLoadBalancingV2TargetGroupTargetDescription
-  { _elasticLoadBalancingV2TargetGroupTargetDescriptionId :: Val Text
+  { _elasticLoadBalancingV2TargetGroupTargetDescriptionAvailabilityZone :: Maybe (Val Text)
+  , _elasticLoadBalancingV2TargetGroupTargetDescriptionId :: Val Text
   , _elasticLoadBalancingV2TargetGroupTargetDescriptionPort :: Maybe (Val Integer)
   } deriving (Show, Eq)
 
@@ -29,13 +30,15 @@ instance ToJSON ElasticLoadBalancingV2TargetGroupTargetDescription where
   toJSON ElasticLoadBalancingV2TargetGroupTargetDescription{..} =
     object $
     catMaybes
-    [ (Just . ("Id",) . toJSON) _elasticLoadBalancingV2TargetGroupTargetDescriptionId
+    [ fmap (("AvailabilityZone",) . toJSON) _elasticLoadBalancingV2TargetGroupTargetDescriptionAvailabilityZone
+    , (Just . ("Id",) . toJSON) _elasticLoadBalancingV2TargetGroupTargetDescriptionId
     , fmap (("Port",) . toJSON . fmap Integer') _elasticLoadBalancingV2TargetGroupTargetDescriptionPort
     ]
 
 instance FromJSON ElasticLoadBalancingV2TargetGroupTargetDescription where
   parseJSON (Object obj) =
     ElasticLoadBalancingV2TargetGroupTargetDescription <$>
+      (obj .:? "AvailabilityZone") <*>
       (obj .: "Id") <*>
       fmap (fmap (fmap unInteger')) (obj .:? "Port")
   parseJSON _ = mempty
@@ -47,9 +50,14 @@ elasticLoadBalancingV2TargetGroupTargetDescription
   -> ElasticLoadBalancingV2TargetGroupTargetDescription
 elasticLoadBalancingV2TargetGroupTargetDescription idarg =
   ElasticLoadBalancingV2TargetGroupTargetDescription
-  { _elasticLoadBalancingV2TargetGroupTargetDescriptionId = idarg
+  { _elasticLoadBalancingV2TargetGroupTargetDescriptionAvailabilityZone = Nothing
+  , _elasticLoadBalancingV2TargetGroupTargetDescriptionId = idarg
   , _elasticLoadBalancingV2TargetGroupTargetDescriptionPort = Nothing
   }
+
+-- | http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticloadbalancingv2-targetgroup-targetdescription.html#cfn-elasticloadbalancingv2-targetgroup-targetdescription-availabilityzone
+elbvtgtdAvailabilityZone :: Lens' ElasticLoadBalancingV2TargetGroupTargetDescription (Maybe (Val Text))
+elbvtgtdAvailabilityZone = lens _elasticLoadBalancingV2TargetGroupTargetDescriptionAvailabilityZone (\s a -> s { _elasticLoadBalancingV2TargetGroupTargetDescriptionAvailabilityZone = a })
 
 -- | http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-elasticloadbalancingv2-targetgroup-targetdescription.html#cfn-elasticloadbalancingv2-targetgroup-targetdescription-id
 elbvtgtdId :: Lens' ElasticLoadBalancingV2TargetGroupTargetDescription (Val Text)