+++ library-gen/Stratosphere/Resources/ElasticBeanstalkApplication.hs
@@ -13,7 +13,7 @@ import Data.Monoid (mempty)
 import Data.Text
 
 import Stratosphere.Values
-
+import Stratosphere.ResourceProperties.ElasticBeanstalkApplicationApplicationResourceLifecycleConfig
 
 -- | Full data type definition for ElasticBeanstalkApplication. See
 -- 'elasticBeanstalkApplication' for a more convenient constructor.
@@ -21,6 +21,7 @@ data ElasticBeanstalkApplication =
   ElasticBeanstalkApplication
   { _elasticBeanstalkApplicationApplicationName :: Maybe (Val Text)
   , _elasticBeanstalkApplicationDescription :: Maybe (Val Text)
+  , _elasticBeanstalkApplicationResourceLifecycleConfig :: Maybe ElasticBeanstalkApplicationApplicationResourceLifecycleConfig
   } deriving (Show, Eq)
 
 instance ToJSON ElasticBeanstalkApplication where
@@ -29,13 +30,15 @@ instance ToJSON ElasticBeanstalkApplication where
     catMaybes
     [ fmap (("ApplicationName",) . toJSON) _elasticBeanstalkApplicationApplicationName
     , fmap (("Description",) . toJSON) _elasticBeanstalkApplicationDescription
+    , fmap (("ResourceLifecycleConfig",) . toJSON) _elasticBeanstalkApplicationResourceLifecycleConfig
     ]
 
 instance FromJSON ElasticBeanstalkApplication where
   parseJSON (Object obj) =
     ElasticBeanstalkApplication <$>
       (obj .:? "ApplicationName") <*>
-      (obj .:? "Description")
+      (obj .:? "Description") <*>
+      (obj .:? "ResourceLifecycleConfig")
   parseJSON _ = mempty
 
 -- | Constructor for 'ElasticBeanstalkApplication' containing required fields
@@ -46,6 +49,7 @@ elasticBeanstalkApplication  =
   ElasticBeanstalkApplication
   { _elasticBeanstalkApplicationApplicationName = Nothing
   , _elasticBeanstalkApplicationDescription = Nothing
+  , _elasticBeanstalkApplicationResourceLifecycleConfig = Nothing
   }
 
 -- | http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-beanstalk.html#cfn-elasticbeanstalk-application-name
@@ -55,3 +59,7 @@ ebaApplicationName = lens _elasticBeanstalkApplicationApplicationName (\s a -> s
 -- | http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-beanstalk.html#cfn-elasticbeanstalk-application-description
 ebaDescription :: Lens' ElasticBeanstalkApplication (Maybe (Val Text))
 ebaDescription = lens _elasticBeanstalkApplicationDescription (\s a -> s { _elasticBeanstalkApplicationDescription = a })
+
+-- | http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-beanstalk.html#cfn-elasticbeanstalk-application-resourcelifecycleconfig
+ebaResourceLifecycleConfig :: Lens' ElasticBeanstalkApplication (Maybe ElasticBeanstalkApplicationApplicationResourceLifecycleConfig)
+ebaResourceLifecycleConfig = lens _elasticBeanstalkApplicationResourceLifecycleConfig (\s a -> s { _elasticBeanstalkApplicationResourceLifecycleConfig = a })